Moving NFT methods to NFT API
